sys-libs/compiler-rt: Migrated patches from the ebuild to the json file
The ebuild executes a script that applies the patches in the json file.
If any of the patches fail, then the script will print to stdout all the
patches that have applied successfully up to the failed patch and
then 'die' will be called. Otherwise, the script will print to stdout
which patches applied successfully and then which patches were not
applicable.

BUG=None
TEST='USE="llvm-next" emerge sys-libs/compiler-rt' and 'emerge
sys-libs/compiler-rt'

Change-Id: I7fb186a0897d28d3c6d89b775ae113191d6227bf