diagnostics: depend on chromeos-dtc-vm not full wilco-bsp
The only real dependency here is on the VM, and not everything else in
the chromeos-bsp-wilco-private package. The chromeos-bsp packages are included
in the chromeos factory install image. The VM is too large for this image
and it will now be brought in as a dependency of the diagnostics package.

BUG=b:138655966
TEST=setup fresh sarien board; build_packages --board=sarien;
build_image --board=sarien; equery-sarien l chromeos-dtc-vm (installed)

Change-Id: I8ed6092b69ba4b6fb4bcf4478b8b25afe2615df7
Reviewed-on: https://chromium-review.googlesource.com/1731680
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Ready: Trent Begin <tbegin@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>