mod_wsgi: fix apr-1-config lookup
The build currently looks for `apr-1-config` in $PATH.
This has been working as we've had apr installed in the SDK,
but since we're dropping that, we notice this misbehavior.
Hack the build env to find the board's config scripts.

BUG=chromium:982351
TEST=SDK & CQ pass

Change-Id: I2683ead04da62fb814584939100cd7e3c436e8ca
Reviewed-on: https://chromium-review.googlesource.com/1730635
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>