chromeos-chrome: Deploy using orderfile for Chrome OS.
We have been working on the tooling for using orderfiles for a while,
and the performance shows 1-3% performance gain to use orderfiles.
Another Chromium CL will set the default path of Chrome OS orderfiles.
The changes to Chrome ebuild will be just adding use flags, and freeze
logic in case some horrible things happen to a specific version of
orderfile.

BUG=chromium:950627
TEST=Tryjob pass https://ci.chromium.org/p/chromeos/builders/general/Try/b8904666195157226576

Change-Id: Iec8e1f557544367c507dc46b39dc5318de8fda78