llvm: Cherry-pick r366466 and r368293 to llvm-next to support some ARM64 relocation types.
LLD failed to resolve R_AARCH64_TLSLE_MOVW_TPREL_G0 relocation type, as reported
in crbug.com/985141. Cherry-pick an upstream fix r368293 on the bug.
r366466 is also cherry-picked because it's depended by r368293 and also implements
several other relocation types.
r366463 is cherry-picked because of dependency, but it's only a refactoring.

BUG=chromium:985141,chromium:986192
TEST=LLVM builds locally. Tryjob passes.

Change-Id: Ibbcfb869e00ee6e3941883727056326d5ab2965c