libchrome: add patch to avoid crashing for mismatched histogram parms.
These mismatches shouldn't happen but crashing is a fairly severe
response for the released products.  Just report the error.

Source code from https://chromium-review.googlesource.com/c/aosp/platform/external/libchrome/+/1767175/

List of CLs combined:
- (r552089) Don't crash for mismatched histogram parms.
https://chromium-review.googlesource.com/c/chromium/src/+/1012469/
- (r545333) Do not record expired histograms.
https://chromium-review.googlesource.com/c/chromium/src/+/957437/
- (r503662) Change Pickle::Write* methods to return void.
https://chromium-review.googlesource.com/c/chromium/src/+/677023/

The first two CLs are what we needed that returns a DummyHistogram
instance instead of a nullptr. We need part of the third CL to align
chromeos aosp repo function definition with that of the first two CLs.

BUG=b:127244647
TEST=emerge-gale libchrome

Change-Id: I47c9daea0ea3dfd5a341a963d2fc8d57f3e619e4