profiles: drop old python version hack
That was written for a system that didn't support parallel python
installs and cross-compilation well.  The Gentoo ecosystem is much
better now, so we shouldn't need this hack anymore.  It also breaks
our ability to build more than one version of python for boards at
a time which is what we want now (Python 2.7 & 3.6).

BUG=chromium:992617
TEST=`emerge-$BOARD dbus-python` produces two sep builds correctly

Change-Id: Ic88b206a4a76a28061667240b46c6b1e2f9c44b4