modemmanager-next: extend the ugly src_test() path hack
We've picked out some of the libraries to run from the sysroot, but
we've missed /lib64 as well as binaries on ${PATH}. Notably, this means
gtester gets picked up from the SDK, but links (LD_LIBRARAY_PATH)
against the sysroot's glib. That doesn't work if, for instance, the
sysroot was using ASAN but the SDK was not.

Extend the LD_LIBRARY_PATH hacks to include both /usr/lib64 and /lib64,
as well as to include /usr/bin (to get the right gtester).

Also drop the unused IUSE=asan. It's only used in the eclasses we
inherit.

BUG=chromium:951639
TEST=asan tryjobs

Change-Id: Iefdc4f5a28e4ed7e04087ffcc6bc6ff217ebeec4