arc-build: Define STRIP/OBJCOPY
On arm/arm64 multilib ARC++ boards, after the install stage,
portage tries to objcopy/strip the binaries, but it uses the
arm version of the toolchain, which is not able to handle
64-bit files:

arm-linux-androideabi-objcopy:/build/kukui/.../vendor/lib64/hw/gralloc.cros.so: File format not recognized
arm-linux-androideabi-objcopy:/build/kukui/.../bin/gralloctest_arm64: File format not recognized
...

Basically, this is because when src_install runs, it runs the
arm64 version, then the arm version, and leaves
CHOST=arm-linux-androideabi behind:
 * abi_arm_64.arm64: running multilib-minimal_abi_src_install
 * CHOST is aarch64-linux-android
 * abi_arm_32.arm: running multilib-minimal_abi_src_install
 * CHOST is arm-linux-androideabi

And then portage just looks for ${CHOST}-strip to strip the
files.

This is also a bug on x86:
 * abi_x86_64.amd64: running multilib-minimal_abi_src_install
 * CHOST is x86_64-linux-android
 * abi_x86_32.x86: running multilib-minimal_abi_src_install
 * CHOST is i686-linux-android

But then i686-linux-android-strip does not exist, and portage
falls back to `strip`, which... well, just works since we're
building on a x86_64 device.

Manually set STRIP/OBJCOPY in the eclass to make sure the right
strip tool is used.

BUG=b:139215445
TEST=emerge-kukui -v arc-cros-gralloc => no error
  file /build/kukui/opt/google/containers/android/vendor/lib*/hw/gralloc.cros.so
  /build/kukui/opt/google/containers/android/vendor/lib64/hw/gralloc.cros.so:
    ELF 64-bit LSB pie executable, ARM aarch64, ..., dynamically linked, stripped
  /build/kukui/opt/google/containers/android/vendor/lib/hw/gralloc.cros.so:
    ELF 32-bit LSB pie executable, ARM, ..., dynamically linked, stripped

Change-Id: Ie886562ee8d5938f8ea3b12da0d65b77c1e97cc9