llvm: Make llvm-tot more directly useful
We've pivoted a bit on how we plan to do LLVM ToT builds. In particular,
we have a script that updates our LLVM_NEXT hash (among other things) to
just point us at the ToT revision to test, rather than opting for having
the ebuild fetch whatever version it wants.

We still want assertions to be enabled/etc. for these ToT builds, so
llvm-tot is still somewhat useful. It just shouldn't be going out of its
way to check out origin/master anymore.

BUG=None
TEST=`USE=llvm-tot emerge llvm`

Change-Id: Ibcd6ed5c9d9da46f42253d361603ccfd5314cb84