python: fix cross-compile logic in sysconfig module
We've long carried hacks in distutils.sysconfig for cross-compiling
Python projects.  The sysconfig module is a large duplication of that
logic which needs the same set of hacks added to it.  We didn't see
this before because the most common user (automake) specifically did
not use the sysconfig module in Python 2, preferring distutils.  With
Python 3, automake defaults to the other module causing us to notice
the missing logic.

BUG=chromium:884766, chromium:982409
TEST=`emerge-arm-generic dbus-python` installs into the right /usr/lib/ dir

Change-Id: I59d1aa0bb9aff07b09ff409e4ee23fbb0ba50ab9