dev-rust/lazy_static: update to 1.4.0
This updates to `spin` 0.5, which fixes an unsoundness in `spin`s
implementation of `RwLock`. This is not used by `lazy_static`, but
fixing the unsoundness now ensures it can't be triggered by unknowing
users in the future.

A few dependencies are compatible with lazy_static 1.x but were pinned
to 1.2.0, so they have been updated to use any 1.x version.

See https://rustsec.org/advisories/RUSTSEC-2019-0013.html for details.

BRANCH=None
TEST=Dependencies still build
BUG=chromium:1002289

Change-Id: Ie33e653688598918d224cf1dcf76e03b821d01c5
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1792665
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>