vp8 encoder: Workaround to force performance mode encode
This is a tradeoff between quality and performance to
save power. Use quality-level two as the default
which will disable a couple of high intensity operations.
For eg: There won't be a 16xMe layer in the hierarhical motion
estimations which should save significant number of cycles.
Also number of pak passes are now limited to 1.

The ideal way to tune the encode quality is by sening the
misc parameter VAEncMiscParameterEncQuality after querying
the driver capability via VAConfigAttribEncQualityRange.
But chromium is currently not sending any misc parameters
and upstream intel-vaapi-driver requires a fix:
https://github.com/intel/intel-vaapi-driver/pull/485
So this is a quick workaround to use the low-power consumption
mode of encode as default for chromeos.

BUG=b:131723850, b:141147405
TEST=encode the video sample using appr.tc and measure the power
consumption. There should be atleast 0.1W power saving.

Change-Id: Ic2c8018d00ded589ed5388367ed2e6e86be4da5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1804119
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>