use hana-pre-cq instead of oak-pre-cq
Oak is just a baseboard and has no build coverage in either the legacy
CQ or the parallel CQ.  This means that the oak build can break in a
persistent way (as it did in https://crbug.com/1003534), and this
breakage will go unnoticed by sheriffs and the rest of cros-oncall.

Instead, use hana-pre-cq.  Hana is an board based on oak and has
coverage both in parallel CQ and by a release builder in the legacy CQ.

BUG=chromium:1003534,chromium:1003827
TEST=Ran a tryjob with hana-pre-cq to verify that it was healthy.

Change-Id: I94da897ce9c699d6c45c1dae9975e1f050c85ff9
Signed-off-by: Ross Zwisler <zwisler@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1804448
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>