dev-lang/rust: upgrade to rust 1.38.0
This change also adds a handful of bare metal thumb targets which can be
used for build cros_ec Rust code. This required a new array of targets
to be handled slightly differently and required one deceptively small
patch, the details of which can be found in "no-test-on-build".

Other than that, this change is fairly standard. A few patches needed
small adjustments to context. The manual download of llvm-project is no
longer needed to make libunwind work as the source distrobution now has
the libunwind code. The upstream patch
"rustbuild-detect-cxx-for-all-targets" is merged in this version of
Rust, so it was also dropped.

TEST=chromiumos-sdk-tryjob
     crosvm's build_test
     build some WIP cros_ec rust code
BUG=None

Cq-Depend: chromium:1826106
Change-Id: I6d6c961fa7faf9b269e18fe392699cde33b3bfaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1830892
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>