gen9_vp9_encoder: Ensure brc-reset irrespective of the frame type
If bitrate control is enabled (CBR & VBR), it is legal to
have a bitrate change/reset even for non-key frame.
Currently the bitrate change is not submitting to the gpu
if it is an inter-prediction frame.
This patch ensures that all internal structures are
getting the updated bitrate values whenever user request.

This patch is a fix for the intel-vaapi-driver upstream bug:
https://github.com/intel/intel-vaapi-driver/issues/473

Here is the intel-vaapi-driver Pull Request:
https://github.com/intel/intel-vaapi-driver/pull/486

BUG=chromium:811912
TEST=Encode a vp9 video with chrome and ensure the bitrate
change in chrome://webrtc-internals

Change-Id: I6d447ec5edfbfe258f197a6df2cfea068d61f373
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1835171
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>