cros-workon: tidy up checkout warning
If people build a cros-workon package but their local checkout is not
at the exact rev it wants, we issue a warning and fallback to cloning
ourselves.  It isn't clear to users what this message means, and the
use of ewarn leads to a lot of post-emerge spam dumped to the console.

Lets switch it to a plain echo and add a little more detail explaining
what is going on.

BUG=None
TEST=emerged a package & saw updated message

Change-Id: I429f9aea660d8f79d14fda674bb715da8e2358b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1838613
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>