tast-use-flags: Add disable_cros_video_decoder flag.
Not all devices support the new video decoders, as they require import mode to
be supported. This CL exposes the disable_cros_video_decoder flag so it can be
used to skip VD tests on devices not supporting the new VDs. Currently we check
whether the device supports Android (which requires import mode), but this check
is not always accurate.

BUG=b:137916185
TEST=emerge tast-use-flags

Change-Id: I56dcb6d7b80f29a0597f9b56d17a9297df372f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1838779
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>