Remove respawn mechanism from the upstart script for rtanalytics.
This respawn mechanism in the Upstart script for mri_package causes more
trouble than its worth and makes it hard to keep our understanding of the
process status on the client side in sync with the actual state of the
process.

We have mechanisms to automatically stop and start the service from the
FE now, which is preferable to an auto-restart.

BUG=chromium:953782
TEST=Verified that respawn does not happen when killing the process.

Change-Id: I203cd473092f0e8c19829588fddf862735de8dc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1841225
Tested-by: Luke Sorenson <lasoren@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>