qualcomm_defconfig: Add CONFIG_DRM_MSM=y to fallback config
As of upstream commit 2eba69071b4b ("drm/msm: Remove Kconfig
default"), DRM_MSM is not longer default to 'y'.  That means we need
to turn it on manually if we want it.  We want it.

This has the side effect of turning on the (super critical) config
option TMPFS since DRM_MSM does "select TMPFS".

NOTE: The fallback config is used when building upstream kernels and
not for anything production.

BUG=None
TEST=Build and boot 5.4-rc2 on cheza

Change-Id: Iddc33570b7b5556272c524b5cba771d66e62d3d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1849971
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Alex Elder <alex.elder@linaro.org>