package.provided: stub out all existing perl packages
The bdepend/revdep rebuild logic is pulling in perl packages into the
sysroot that shouldn't.  While we figure that out, lets disable all
perl packages for the board.  We don't actually care about perl there
so it's not a problem, but this gives us space to figure out the bug.

BUG=chromium:1014105
TEST=CQ passes

Change-Id: Idc1b7c35c816174d72783b53bc527a202548a07d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1861355
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>