Add sys-apps/install-xattr to SDK system set
Builders are racing with this package being installed to the SDK root
and segfaulting when they try to call the binary that is in the process
of being replaced. Add it to the system set so that any operations
involving this package happen before any other packages try to build.

BUG=chromium:1015092, chromium:1015026
TEST=None

Change-Id: I4ae17c76604321973df11e3fbf42c1bf216cf697
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1865800
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>