dev-rust/protobuf: add version 2.8.1
This version of rust-protobuf generates code that does not trigger the
warning about missing `dyn` on trait objects.

The SLOT of all of these packages is transitioned from "${PV}/${PR}" to
"0/${PVR}" in order to prevent file collisions with parallel
installations; in particular, dev-rust/protobuf-codegen installs a
binary, which will conflict with other versions of this package. The
previous version of each package is blocked in order to cause the old
version to be uninstalled.

All users of these packages are also updated to use the exact same
DEPEND syntax for consistency and ease of searching, matching the style
previously used in crostini_client.

BUG=chromium:1015571
TEST=emerge-nami dev-rust/protoc-rust
TEST=sudo emerge dev-rust/protobuf-codegen

Change-Id: Ic46d028d6d4a609ecc0213413bd2d23eb501a898
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1869744
Reviewed-by: Stephen Barber <smbarber@chromium.org>