squashfs-tools: Pick race condition fix from upstream
This grabs a patch found in the squashfs git repo on kernel.org, AKA:

  git://git.kernel.org/pub/scm/fs/squashfs/squashfs-tools.git

It's postulated that this could fix race conditions we're seeing in
the linked bug, specifically when you see this message:

  writer: Lseek on destination failed because Bad file descriptor, offset=...

The important part here is the "Bad file descriptor".  If we were
truly out of space (like the next error that is printed postulates) we
would have had a different errno.

...even more telling is when that error is printed while the stats are
shown, which (I believe) is after the "fd" is closed by the main
thread.

I can't reproduce the problem seen on the builders, but let's pick
this fix and see if the problem goes away.  If it's at all
interesting, it seems that Ubuntu picked this same fix against their
squashfs-tools-4.3:

  https://launchpad.net/ubuntu/+source/squashfs-tools/1:4.3-4ubuntu1

At some point it might be interesting to just switch to an upstream
squashfs-tools, but I will leave that as an exercise for people who
are more invested in this tool.

BUG=chromium:1015117
TEST=emerge squashfs-tools and confirm demo_resources still builds.

Change-Id: I1dae0c3bb5f5c5beb6cd8ed1d02296c51f73dbed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1869405
Tested-by: Douglas Anderson <dianders@chromium.org>
Auto-Submit: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>