cros-rust: Remove the crate from the registry prior to install.
When a rust crate is being installed, cargo should not be aware of the
crate to avoid reading the crate while it is in an inconsistent state.
This is achieved by making sure the symlink to the installed crate
doesn't exist while the crate is being extracted.

BUG=chromium:999931
TEST=Dry run passes

Change-Id: I1293e4b8177719493bd783033e2b252ad0372c39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1874033
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>