qualcomm_defconfig: Add CONFIG_TMPFS=y explicitly
Normally a 'defconfig' doesn't have entries that are default or
selected by another config that's already on.

Right now CONFIG_TMPFS=y is selected by 'CONFIG_DRM_MSM=y' so
technically it doesn't belong in the defconfig.

...but the defconfig we have here isn't truly a defconfig.  It's a
config that we want to work across whatever random kernel version
someone is building against.  That means it's probably a good idea to
be explicit and specify that we need TMPFS since the failure mode
without it is so bad.

The motivation here is that when DRM_MSM stopped being default on then
all of a sudden nothing was turning TMPFS on.  This took a while to
debug.

BUG=None
TEST=None

Change-Id: If53d58c87130987ec2fda8b197228d63feb3f4ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1878029
Tested-by: Douglas Anderson <dianders@chromium.org>
Auto-Submit: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>