dev-rust: use LICENSE=metapackage for empty crates
These things don't install any files, so setting LICENSE doesn't make
sense, and it's extremely unlikely that if we installed the upstream
code, that BSD-Google would be the correct license.

Use LICENSE=metapackage which is intended for empty ebuilds that don't
install files.

Update the eclass to enforce this logic.

BUG=chromium:348750
TEST=CQ passes

Change-Id: Iaf50c633ff35a7982ba453fcc504a19497a6b517
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1890539
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>