amd64: Enable U-Boot and Tianocore on all 64-bit x86 boards
These two alternative firmware options mostly work on these x86 boards. We
don't actually have test coverage on this and it is a volunteer effort.
But by enabling these we can at least get better than the current
behaviour on most platforms, which is to not have any alternative firmware
or legacy mode.

Enable these two on all 64-bit x86 boards.

BUG=chromium:998527
TEST=./build_packages --board=nami
See that it builds U-Boot and Tianocore

Change-Id: I6a0537663583fe74bb1abd94dd71e2b435d378cf
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1894203
Reviewed-by: Mike Frysinger <vapier@chromium.org>