chromite-sdk: split out from main chromite package
We aren't installing chromeos-base/chromite into the SDK currently,
so the use of IUSE=cros_host there is just confusing.  Since the
only reason we want things in the SDK at this point are for python
symlinks that rarely change, create a new package to hold the logic
so we aren't constantly updating the SDK for unrelated chromite
source changes.

BUG=chromium:981405
TEST=CQ passes

Change-Id: I3c8b58a3c50d9747e9c9bdae141465fa2588b3a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1898672
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>