autotest: Stop deploying video_HangoutHardwarePerf autotest.
This CL stops the video_HangoutHardwarePerf autotest from deploying. This test
is currently not running anymore as it was replaced by the DecodeEncodeAccelPerf
Tast test, but it was still being used by the video_PowerConsumption test. CL
crrev.com/c/1902559 removed this test, so it's now safe to remove the
video_HangoutHardwarePerf test.

BUG=chromium:1021472
TEST=build autotest-chrome

Cq-Depend: 1902559
Change-Id: I9ef6173f8f2fbbdcc468d5ba634ad89747ef1438
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1902560
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>