chromeos-base/autotest-server-tests-shill: drop silly dependencies
This CL was quite wrong:
  https://chromium-review.googlesource.com/272531
We're not actually using hostapd for anything on the target (at build-
or run-time) -- the key point was that the target *router* needs a newer
hostapd, but that's totally not what DEPEND means here.

Drop it, so we stop building hostapd needlessly.

Also, DEPEND is not the right place for build blockers. This is an
ancient rule anyway, so it shouldn't be relevant anyway. Let's drop the
bad example.

BUG=none
TEST=none

Change-Id: Iccf6de2965c88ac86d1684842734ae4c5e6c7922
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1906978
Tested-by: Brian Norris <briannorris@chromium.org>
Auto-Submit: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>