coreboot: Use amd-firmware dependency only for grunt
The AGESA firmware for stoney was being built as a requirement of the
coreboot build.  We need to remove it from there since we're using a
different ebuild for the picasso AGESA firmware.  It's getting built as
a part of the chipset-stnyridge-private/coreboot-private-files ebuild
now.

BUG=b:144238104
TEST=emerge-grunt --unmerge amd-firmware; \
emerge-grunt coreboot

Cq-Depend: chrome-internal:2126712
Change-Id: I80a071a64e788902f06e3dc365281bc529be8dc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1909389
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Auto-Submit: Martin Roth <martinroth@chromium.org>