cr-Os: add -Os opt-outs to hot packages; update EAPIs
The EAPI updates are limited to media-libs/speex,
media-sound/adhd, and sys-devel/binutils. They're intended to make `repo
upload` happy. :)

Since the world is currently on -O2, this is intended to be a functional
nop.

That aside, the top 25 hottest packages in Chrome OS are:
[1] "chromeos-base/chromeos-chrome": 8,801,051,649 samples (39.63%; cumulatively, 39.63%)
[2] "sys-kernel/chromeos-kernel-4_19": 5,584,794,650 samples (25.15%; cumulatively, 64.78%)
[3] "sys-kernel/chromeos-kernel-4_4": 5,584,794,650 samples (25.15%; cumulatively, 89.93%)
[4] "sys-libs/glibc": 903,941,024 samples (4.07%; cumulatively, 94.00%)
[5] "media-libs/mesa": 374,902,951 samples (1.69%; cumulatively, 95.69%)
[6] "sys-apps/dbus": 117,691,323 samples (0.53%; cumulatively, 96.22%)
[7] "sys-libs/libcxx": 84,404,884 samples (0.38%; cumulatively, 96.60%)
[9] "media-libs/intel-hybrid-driver": 76,872,383 samples (0.35%; cumulatively, 97.33%)
[10] "media-sound/adhd": 74,646,397 samples (0.34%; cumulatively, 97.66%)
[12] "media-libs/cmrt": 60,655,575 samples (0.27%; cumulatively, 98.25%)
[13] "net-dns/avahi": 45,851,567 samples (0.21%; cumulatively, 98.46%)
[14] "sys-devel/binutils": 40,366,343 samples (0.18%; cumulatively, 98.64%)
[15] "chromeos-base/libchrome": 34,296,840 samples (0.15%; cumulatively, 98.80%)
[16] "sys-apps/journald": 27,168,385 samples (0.12%; cumulatively, 98.92%)
[17] "chromeos-base/gestures": 24,560,365 samples (0.11%; cumulatively, 99.03%)
[18] "media-libs/fontconfig": 23,202,205 samples (0.10%; cumulatively, 99.13%)
[19] "sys-fs/udev": 19,625,596 samples (0.09%; cumulatively, 99.22%)
[20] "dev-libs/nss": 17,701,444 samples (0.08%; cumulatively, 99.30%)
[21] "dev-libs/openssl": 14,282,042 samples (0.06%; cumulatively, 99.37%)
[22] "x11-libs/libdrm": 12,871,031 samples (0.06%; cumulatively, 99.42%)
[23] "chromeos-base/shill": 11,436,056 samples (0.05%; cumulatively, 99.48%)
[25] "media-libs/speex": 7,261,191 samples (0.03%; cumulatively, 99.55%)

This CL adds a primitive to opt packages into -O2, and applies it to all
packages above that make use of -O* CFLAGS provided by the system.

It's likely that `grep cros_optimize_package_for_speed` can simplify
some of our tooling that reasons about all of this if we apply it
consistently, so I leave commented out versions of that in ebuilds that
try to control their own optimization level destiny.

"Top 25" was an arbitrary selection. Mostly I'm keying off of that
covering 99.5% of samples as a measure of "anything below this line
probably doesn't matter". If we want to instead target 99% or similar,
that's cool, too.

Astute readers may note that some numbers are missing above. These are
packages present in other overlays.

BUG=chromium:908661
TEST=Tryjobs for correctness; previous perf tests for where to set
     thresholds/etc.

Change-Id: Ibc1750a31f65cf119f056e90c352ca8286a37aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1912822
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>