dev-go/cdp: local revert to fix goroutine leak
The main patch this change reverts breaks the exit condition checking
for the session.Manager goroutine in certain failure conditions because
it propagates the original error that caused the failure. When this
happens, the goroutine ends up spinning and consuming a significant
amount of CPU, which can interfere with subsequent tests.

BUG=chromium:1019277
TEST=no failures on Tast

Change-Id: I3b83e39df3e41a75e93f093a6958a785ce2b14ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1914898
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>