libva-intel-driver, intel-hybrid-driver: Remove blitter usage
See CVE-2019-0155. libva uses the blitter to clear the VASurface before
decoding into it. The VA API allows for the video to be decoded into a
subrectangle of the surface, in which case the rest of the surface gets
cleared to a solid color.  Chrome always decodes to the entire surface
and the clear can be removed since the clear color is fully overwritten.

BUG=chromium:1024182
TEST=Video decode works for GPUs and codecs affected

Change-Id: I503e084a951a2fe2084b091c3a0f3b127e59e8eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1923067
Tested-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>