arc-llvm: Disable unused features
arc-mesa is the only current consumer of arc-llvm. It doesn't use rtti
or exception handling, so we can disable both.  This saves a couple of
MBs on the vendor image.

To make sure the host llvm-config BuildVariables.inc matches this and
future changes, sync the whole thing from the target instead of just
replacing the static library flags.

BUG=b:112313068
TEST=build grunt and betty-arc64

Change-Id: I025193d440b3c1bb1b784ca799b89b1084592efe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1930187
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>