iwlwifi_rescan: Issue SBR (secondary bus reset) before rescan
Adding a new set of rules to log the PCIe root port for WiFi devices and
issue SBR on the root port in case we need to "revive" the device - we
must warm-reset it.

The workaround of rescanning the PCI bus, the way we implement our
workaround is incomplete. I would even say that it isn't within the
spec. I will try to summarize (sorry for the length).

Diving into PCI express specification
(http://www.lttconn.com/res/lttconn/pdres/201402/20140218105502619.pdf
- page 506) there are two types of resets - conventional (cold, warm,
and hot) reset and FLR (function-level reset).

by removing the device - we are not triggering an FLR and we are not
triggering any of the conventional resets. We need to do a "warm"
conventional reset before we rescan, otherwise, as far as I understand,
the behavior is inconsistent. The only way to tiger a warm reset is by
setpci to the bridge control register (bit 6) forcing a secondary bus
reset (https://cds.cern.ch/record/551427/files/cer-2308933.pdf - page 52).

BUG=b:145604778
TEST=build kefka-image, check dmesg, make sure workaround works

Change-Id: I0e4717031312e703783d1e3f5ae100ec2cb79798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1951305
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Alex Levin <levinale@chromium.org>
Commit-Queue: Alex Levin <levinale@chromium.org>