dev-db/mariadb: Add hack to force building after openssl
The mariadb build doesn't depend on openssl, but picks up the SDK's
openssl when cmake detects installed system libraries. That breaks the
build later when the compiler can't find the OpenSSL headers in the
board sysroot.

As a workaround, add an explicit dependency on openssl for now.

BUG=chromium:1031533
TEST=emerge-$BOARD --unmerge openssl; emerge-$BOARD -v1 mariadb

Change-Id: I7f2edea6ab7773a0b9cf1aad48558360810542ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1955546
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>