ec-boards: remove boards which are not tested by buildall
The skip_boards list is getting more boards, but our commit queue only
runs on a limited set of boards. This means the commit queue may
submit changes which could break certain builders.

Synchronize this list to only build EC images which are tested by
buildall for now. Plan is to later add a check that none of the
skipped boards are in this list.

BUG=chromium:1031621
TEST=emerge-glados chromeos-ec succeeds

Cq-Depend: chrome-internal:2306522,chrome-internal:2306263,chromium:1963073
Change-Id: If8dc5e755151762d309a49834e3f2d328a37da04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1962820
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>