Add support for non-virtual chromeos-config-bsp package
This CL adds a dependency on chromeos-config-bsp behind the new USE flag
`generated_cros_config` to all packages that previously depended on
chromeos-config. This USE flag is only enabled on overlays that have
been migrated to a new schema for chromeos-config-bsp, as this package
name does not exist yet for most boards.

The purpose of this new dependency is to bypass the configuration
generation step of chromeos-config and the problematic virtual package
that was previously used by overlays to drag in config-bsp packages on
an overridable basis. Packages that use this new USE flag now DEPEND
directly on the package providing the configuration, which means that
changes in the subslot of that configuration package will properly
trigger rebuilds of packages that this CL changes.

BUG=chromium:1016803
TEST=None

Change-Id: Ia8a66126790d74e8700d161308c45d00de85b300
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1966213
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>