bootchart: include default-disabled bootchart in test image.
This includes but disables bootchart in the the test image by default.
Enablement of bootchart is controlled by the presence of
"cros_bootchart" in kernel args. The boot perf test modifies kernel
args before and after the test to get bootchart data of the test.

BUG=chromium:783065
TEST=test_that $dut BootPerfServerCrosPerf and check /var/log/bootchart/
on DUT

Cq-Depend: chromium:1966666
Change-Id: Iae7d0318d89e0b34074c891e4282d785393cff03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1966861
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>