chromeos-bootimage: Use the special zork coreboot
When building on zork we must use the special coreboot ebuild so we get
the right source. Add a special case for now.

It does not seem to be worth making this into a 'virtual' ebuild since
these changes will be dropped, hopefully in a few weeks. The change to
virtual would be quite intrusive and harder to back out.

BUG=chromium:1034094
TEST=cros_workon-zork start coreboot-zork
emerge-zork chromeos-bootimage

Change-Id: Ie04fdffcc1a650e0a1f459cba53edb4155ac2049
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1970197
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>