edk2: Drop the coreboot dependency
With zork we currently use a different coreboot ebuild so we don't want to
depend on coreboot.

In any case the real dependency is coreboot-sdk apparently, so update to
EAPI 7 and use BDEPEND.

Also fix one of the patches that is in the wrong format, so that emerge
suceeds.

BUG=chromium:1034094
TEST=cros_workon-zork start coreboot-zork
emerge-zork chromeos-bootimage

Change-Id: I35b91e8d0cf0018d63df1c060755ecb2581e3bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1971012
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>